Reimplemented Css CLass, added cache_disabled_urls, and now we test_cache
Cache is now being tested via a webserver with Redis. Users can submit urls that should not be cached for Feed or blog post scraping where you only wish to catch new pages and ignore old ones. A new CSS class has been added with backwards compatibility in mind accept for jquery like css functions for that we still have the old code which can be imported as Pyq.